Final placeholders before submitting to the AMS
Working off work by @jjhelmus and @dopplershift I filled out the remianing
sections. The second page is pretty much a copy and paste. In the end this will
hold information for attendees.

I filled in most data from the AWS perspective.. We will need to flesh that out
some more in the future.